manage.py: finish certificate creation, then send mail
move sending of EMail to user after the certificate is created and this
status is stored to db. This keeps sync between the certificates created
by EasyRSA and out database, to prevent double creation of certs.
The db-field is name "generation-date" so it should only reflect when the
cert was created / request was managed.
Sending emails is very different and should be tracked in it'S own field,
if required.